feat: honk flows exposed through wasm (#6096)
Adds the Honk prove, verify, and store_vk flows for WASM (through
bb.js).

Closes https://github.com/AztecProtocol/barretenberg/issues/970.

+++ barretenberg/cpp/src/barretenberg/eccvm/eccvm_verifier.cpp
@@ -123,7 +123,7 @@ bool ECCVMVerifier::verify_proof(const HonkProof& proof)
     const size_t log_circuit_size = numeric::get_msb(circuit_size);
     auto sumcheck = SumcheckVerifier<Flavor>(log_circuit_size, transcript);
     FF alpha = transcript->template get_challenge<FF>("Sumcheck:alpha");
-    std::vector<FF> gate_challenges(numeric::get_msb(key->circuit_size));
+    std::vector<FF> gate_challenges(static_cast<size_t>(numeric::get_msb(key->circuit_size)));
     for (size_t idx = 0; idx < gate_challenges.size(); idx++) {
         gate_challenges[idx] = transcript->template get_challenge<FF>("Sumcheck:gate_challenge_" + std::to_string(idx));
     }
